requirements.txt
allure-behave
allure-python-commons
Appium-Python-Client
behave


